Split classes and added some docs
