+++ Components/ModInfo.cs
@@ -150,10 +150,5 @@ namespace tModloaderDiscordBot.Components
 		}
 	}
 
-	public class VoteData
-	{
-		public int score;
-		public int votes_up;
-		public int votes_down;
-	}
+
 }
