+++ Components/SiteStatusCode.cs
@@ -1,0 +1,7 @@
+	public enum SiteStatusCode
+	{
+		Offline,
+		Online,
+		Unknown,
+		Invalid
+	}
