+++ Components/VoteData.cs
@@ -1,0 +1,6 @@
+	public class VoteData
+	{
+		public int score;
+		public int votes_up;
+		public int votes_down;
+	}
